+++ app/Slide.php
@@ -7,10 +7,11 @@ namespace App;
 use Illuminate\Support\Facades\File;
 use Illuminate\Database\Eloquent\Model;
 
-class Slide extends Model
+class Slide extends HeaderImageModel
 {
 
     const ARTWORKS_DIR = '/images/intro/';
+    const IMAGE_PROPERTY='image';
 
     protected $fillable = [
                 'title',
@@ -37,35 +38,5 @@ class Slide extends Model
     public function scopePublished($query)
     {
         return $query->where('publish', '=', 1);
-    }
-
-    public function getImagePathAttribute()
-    {
-        return asset(self::ARTWORKS_DIR . '/' . $this->id . '/' . $this->image . '.jpg');
-    }
-
-    public function getImageSrcsetAttribute()
-    {   
-        return 
-        asset(self::ARTWORKS_DIR . '/' . $this->id . '/' . $this->image . '.jpg'). ' 400w, ' .
-        asset(self::ARTWORKS_DIR . '/' . $this->id . '/' . $this->image . '.800.jpg'). ' 800w, ' .
-        asset(self::ARTWORKS_DIR . '/' . $this->id . '/' . $this->image . '.1200.jpg'). ' 1200w, ' .
-        asset(self::ARTWORKS_DIR . '/' . $this->id . '/' . $this->image . '.original.jpg'). ' 1201w' ;
-    }
-
-    public function getPath($create = false)
-    {
-        $folder_name = $this->id;
-        $path = public_path() .  self::ARTWORKS_DIR . $folder_name . '/';
-        if (!File::exists($path) && $create) {
-            File::makeDirectory($path);
-        }
-        return $path;
-    }
-
-    public function removeImage()
-    {
-        $dir = $this->getPath();
-        return File::cleanDirectory($dir);
     }
 }