+++ database/migrations/2020_10_01_112350_update_collections_main_images.php
@@ -1,0 +1,28 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class UpdateCollectionsMainImages extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+		  DB::statement('UPDATE `slides` SET `image` =  CONCAT(id,"/",image)');
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+		  DB::statement('UPDATE `slides` SET `image` =  REPLACE(image,CONCAT(id ,"/"),"")');
+    }
+}