+++ resources/views/collections/form.blade.php
@@ -90,10 +90,13 @@
 		{!! Form::file('main_image', array('class' => 'form-control', 'placeholder' => '#666666'))!!}
 		<p>obrazok bude automaticky zmenseny na sirku 1400px</p>
 		<p>šírka min: 1400px<br>formát: JPG (vysoka kompresia ~50-60%)</p>
-		@if (isset($collection) && $collection->getHeaderImage())
+		@if (isset($collection) && $collection->header_image_src)
 			<div class="primary-image">
 				<b>Aktuálny obrázok:</b>
-				<img src="{{ $collection->getHeaderImage() }}" class="img-responsive">
+				<img class="img-responsive"
+					src="{{ $collection->header_image_src }}" 
+					 srcset="{{ $collection->header_image_srcset }}" 
+					 onerror="this.onerror=null;this.srcset=''"/>
 			</div>
 		@endif
 	</div>